Chroma/requirements.txt
speedtest-cli
requests
